requirements.txt
# tensorflow
numpy
pandas
scikit-learn
Pillow
flask
pyzbar
pytesseract
opencv-python
requests
ultralytics